+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 * Added some minor rendering optimizations.
 * Fixed EIcon blocking (right-click on an EIcon to block it)
+* Added a configuration setting to have a PM tab open when someone starts typing to you
+  (instead of waiting until they actually send you a message).  You can also control whether
+  this pings you.